+++ tests/testthat/test-use_project_directory.R
@@ -25,5 +25,6 @@ test_that("use_project_directory works", {
   expect_true(fs::file_exists("_plan.R"))
   expect_true(fs::file_exists("_config.R"))
   expect_true(fs::file_exists(".gitignore"))
+  expect_true(fs::file_exists(".Renviron"))
   expect_true(fs::file_exists("air.toml"))
 })