+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Revise targets plan template
 
+* Use environment variables to specify parallel processing in targets pipelines
+
 # projectsetup 1.0.0
 
 * Initial version of `projectsetup`, forked from `uo-cmor/CMORprojects`